services/auth/requirements.txt
Flask==1.1.2
requests~=2.25.0
waitress==1.4.4
Flask_Cors==3.0.9
python-dotenv==0.15.0
